LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/path-objects/2d.path.isPointInPath.basic.2-expected.txt

2d.path.isPointInPath.basic.2
isPointInPath() detects whether the point is inside the path
Actual output:

PASS isPointInPath() detects whether the point is inside the path
